feat(项目): 完成项目代码编写

+++ internal/internal/config/retry.go
@@ -11,7 +11,7 @@ type Retry struct {
 	// 间隔
 	Interval time.Duration `default:"5s" json:"interval,omitempty" yaml:"interval" xml:"interval" toml:"interval"`
 	// 最大次数
-	Times uint8 `default:"10" json:"times,omitempty" yaml:"times" xml:"times" toml:"times"`
+	Times uint32 `default:"10" json:"times,omitempty" yaml:"times" xml:"times" toml:"times"`
 	// 每次重试拉取个数
 	Count int `default:"10" json:"count,omitempty" yaml:"count" xml:"count" toml:"count"`
 	// 最长执行时间
