+++ internal/internal/repository/internal/core/task.go
@@ -11,7 +11,7 @@ type Task interface {
 
 	Get(task *model.Task, columns ...string) (bool, error)
 
-	GetsRunnable(count int, times uint8, maximum time.Duration, excludes ...*model.Task) (*[]*model.Task, error)
+	GetsRunnable(count int, times uint32, maximum time.Duration, excludes ...*model.Task) (*[]*model.Task, error)
 
 	Update(task *model.Task, columns ...string) (int64, error)
